Add files via upload
The bicycle sales company wants to determine the purchasing patterns of different demographic groups to identify areas where it needs to concentrate its marketing efforts. The company also wants to see if there are purchasing differences between the countries where it operates.
The aim was to determine where marketing effort needs to be applied in specific markets to reach under-represented demographic groups and to see if there are differences between the countries where it operates.
The chart graphically shows that the youth age group is globally the poorest area of sales. Female adults are buying the most products.